Add files via upload
In a file called Nim.java, there is  main method to play the following version of the game of Nim:

There are also some mix basic functions in main class like fizz buzz  etc
The user will play against the computer. A random number decides who takes first turn.
A random number between 10 and 100 inclusive represents the number of matchsticks on the table.
The players take turns to pick up a number of matchsticks:
When it is the user's turn, the programs asks the user how many matchsticks s/he wishes to pick up.
When it is the computer's turn, a random number determines how many matchsticks it will pick up.
However, they must pick up at least one and at most half of the matchsticks. For example, if there are 16 matchsticks, they can pick up between 1 and 8 inclusive; if there are 15 matchsticks, they can pick up between 1 and 7 inclusive.
The player who picks up the last remaining matchstick loses.
Include enough print/println statements so that the user knows what's going on.